+++ src/renderers.js
@@ -5,6 +5,7 @@ const mapboxGlOption = { name: 'Mapbox GL JS', value: 'mapbox-gl' };
 const glVectorRenderers = [
   mapboxGlOption,
   { name: 'Maplibre GL JS', value: 'maplibre-gl' },
+  { name: 'MapTiler SDK', value: 'maptiler-sdk' },
 ];
 
 const typeToRenderers = {
@@ -14,6 +15,7 @@ const typeToRenderers = {
   'vector-gl': glVectorRenderers,
   'mapbox-gl': glVectorRenderers,
   'maplibre-gl': glVectorRenderers,
+  'maptiler-sdk': glVectorRenderers,
 };
 
 export const getRenderers = (map, sources) => {
@@ -22,6 +24,8 @@ export const getRenderers = (map, sources) => {
   if (map.type !== 'mapbox-gl') return typeToRenderers[map.type];
 
   // If using the mapbox:// protocol, force mapbox-gl
+  // TODO instead of forcing this, we should consider replacing all necessary
+  // mapbox protocols in sprites and glyphs with full URL
   if (map.url && isMapboxUrl(map.url)) {
     mapboxOnly = true;
   } else if (sources) {